requirements.txt
pyshark==0.6
scapy==2.5.0
psutil==5.9.8
PyYAML==6.0.1
cryptography==42.0.5
